ЛР4/requirements.txt
pydantic
tk
pycryptodome